Change the start API
This way starting the API, the indexer/parser or both is completely
agnostic of the underlying OS. In addition, the `bin` folder was removed
as the concept of binaries does not match with a backend service like
this.
