--- bin/api
@@ -1,13 +1,0 @@
-#! /usr/bin/env node
-
-'use strict'
-
-const config = require('config')
-
-if (config.newRelic.licenseKey) {
-  require('newrelic')
-}
-
-const api = require('../src/api')
-
-api.start()
