--- bin/indexer
@@ -1,13 +1,0 @@
-#! /usr/bin/env node
-
-'use strict'
-
-const config = require('config')
-
-if (config.newRelic.licenseKey) {
-  require('newrelic')
-}
-
-const indexer = require('../src/indexer')
-
-indexer.start()
