+++ package.json
@@ -7,12 +7,12 @@
   "main": "server.js",
   "repository": "autonomoussoftware/ethereum-blockchain-indexer",
   "scripts": {
-    "api": "./bin/api",
+    "api": "node ./src/api",
     "check": "dependency-check . && dependency-check . --unused --no-dev",
     "check:outdated": "npm-check -p -i husky -i eslint-config-plugin:*",
     "coverage": "nyc --lines 95 --functions 95 --branches 95 --reporter=lcov --reporter=text npm test",
     "dev": "nodemon -w .",
-    "indexer": "./bin/indexer",
+    "indexer": "node ./src/indexer",
     "lint": "eslint --cache .",
     "precommit": "npm run lint",
     "prepush": "npm run check && npm run test",
